backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
pytest==8.3.3
pytest-asyncio==0.24.0
